ref: topology tests helpers
